VocabLib/include/DirectIndex.hpp VocabLib/src/DirectIndex.cpp VocabLib/tests/DirectIndex_test.cpp * Moved direct index implementation to a class.
VocabLib/include/VocabTree.h
* Refactor due to change in direct index implementation.

VocabLib/tests/VocabTree_test.cpp
* Minor change, changed loop variable type.

+++ VocabLib/tests/VocabTree_test.cpp
@@ -149,7 +149,7 @@ TEST(VocabTree, TestDatabase) {
 	for (size_t imgIdx = 0; imgIdx < keysFilenames.size(); imgIdx++) {
 		db->getDbBoWVector(imgIdx, dbBowVector);
 		ASSERT_TRUE(dbBowVector.rows == 1);
-		for (size_t i = 0; i < dbBowVector.cols; i++) {
+		for (int i = 0; i < dbBowVector.cols; i++) {
 			ASSERT_TRUE(dbBowVector.at<float>(0, i) >= 0);
 			ASSERT_TRUE(dbBowVector.at<float>(0, i) <= 1);
 		}
